// src/utils/enums.ts
export enum GetstreamActionsTypesNames {
  CreateUser = "CREATE-USER",
  UpdateUser = "UPDATE-USER",
  DeleteUser = "DELETE-USER",

  CreateGuild = "CREATE-GUILD",
  UpdateGuild = "UPDATE-GUILD",
  DeleteGuild = "DELETE-GUILD",

  CreateLobby = "CREATE-LOBBY",
  UpdateLobby = "UPDATE-LOBBY",
  DeleteLobby = "DELETE-LOBBY",

  CreateGuildUser = "CREATE-GUILD-USER",
  DeleteGuildUser = "DELETE-GUILD-USER",

  CreateLobbyUser = "CREATE-LOBBY-USER",
  DeleteLobbyUser = "DELETE-LOBBY-USER",

  DeleteGetStreamChannel = "DELETE-GETSTREAM-CHANNEL",
  CreateGetStreamChannel = "CREATE-GETSTREAM-CHANNEL",
  AddUserToGetStreamChannel = "ADD-USER-TO-GETSTREAM-CHANNEL",
}